Return integer BL quantities

+++ backend/IQBF.Application/DTOs/BLs/BLDto.cs
@@ -1,2 +1,2 @@
 namespace IQBF.Application.DTOs.BLs;
-public sealed record BLDto(Guid Id, string Code, decimal TotalQuantity, bool IsActive, Guid ShipId, string ShipName, Guid ProductId, string ProductName);
+public sealed record BLDto(Guid Id, string Code, int TotalQuantity, bool IsActive, Guid ShipId, string ShipName, Guid ProductId, string ProductName);
